+++ gtest/catalog.cc
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
 
 #include "../src/registry.hpp"
+#include "options.hpp"
 #include "../src/catalog.hpp"
 
 class catalog : public ::testing::Test
@@ -93,18 +94,71 @@ Step h
       step([](const cuke::value_array&, const auto&, const auto&) {}, "i",
            step::type::when));
 
-  cuke::registry().sort_steps_by_type();
-
   EXPECT_EQ(expected, cuke::catalog::as_readable_text());
 }
 
-TEST_F(catalog, json_init)
+TEST_F(catalog, as_json_1)
 {
+  const std::string expected =
+      R"({
+  "steps_catalog": [
+    {
+      "definition": "a step with {int} and {string}",
+      "type": "Step",
+      "var_types": [
+        "int",
+        "string"
+      ]
+    },
+    {
+      "definition": "another step with {word} and {}",
+      "type": "Step",
+      "var_types": [
+        "word",
+        "anonymous"
+      ]
+    }
+  ]
+})";
+
   cuke::registry().push_step(cuke::internal::step(
       [](const cuke::value_array& values, const auto&, const auto&) {},
       "a step with {int} and {string}"));
   cuke::registry().push_step(cuke::internal::step(
       [](const cuke::value_array& values, const auto&, const auto&) {},
       "another step with {word} and {}"));
-  std::cout << cuke::catalog::as_json() << std::endl;
+
+  EXPECT_EQ(expected, cuke::catalog::as_json());
+}
+
+TEST_F(catalog, as_json_2)
+{
+  using namespace cuke::internal;
+
+  const std::string expected =
+      R"({"steps_catalog":[{"definition":"lets start with {}, {double} and {int}","type":"Given","var_types":["anonymous","double","int"]},{"definition":"lets start with {float} and {double}","type":"Given","var_types":["float","double"]},{"definition":"something with {string} and {long} happens","type":"When","var_types":["string","long"]},{"definition":"{float} added to {double}","type":"When","var_types":["float","double"]},{"definition":"step with {byte} and {word}","type":"Then","var_types":["byte","word"]},{"definition":"we have a lot {},{double},{int},{word},{string},{string},{int},{int},{string},{float},{byte}","type":"Then","var_types":["anonymous","double","int","word","string","string","int","int","string","float","byte"]}]})";
+
+  cuke::registry().push_step(cuke::internal::step(
+      [](const cuke::value_array& values, const auto&, const auto&) {},
+      "step with {byte} and {word}", step::type::then));
+  cuke::registry().push_step(cuke::internal::step(
+      [](const cuke::value_array& values, const auto&, const auto&) {},
+      "something with {string} and {long} happens", step::type::when));
+  cuke::registry().push_step(cuke::internal::step(
+      [](const cuke::value_array& values, const auto&, const auto&) {},
+      "lets start with {}, {double} and {int}", step::type::given));
+  cuke::registry().push_step(cuke::internal::step(
+      [](const cuke::value_array& values, const auto&, const auto&) {},
+      "lets start with {float} and {double}", step::type::given));
+  cuke::registry().push_step(cuke::internal::step(
+      [](const cuke::value_array& values, const auto&, const auto&) {},
+      "{float} added to {double}", step::type::when));
+  cuke::registry().push_step(cuke::internal::step(
+      [](const cuke::value_array& values, const auto&, const auto&) {},
+      "we have a lot "
+      "{},{double},{int},{word},{string},{string},{int},{int},{string},{float},"
+      "{byte}",
+      step::type::then));
+
+  EXPECT_EQ(expected, cuke::catalog::as_json(-1));
 }