+++ src/catalog.cpp
@@ -1,13 +1,19 @@
 #include "catalog.hpp"
 
+#include "options.hpp"
 #include "registry.hpp"
+#include "util.hpp"
 
+#ifdef HAS_JSON
 #include <nlohmann/json.hpp>
+#endif  // HAS_JSON
 
 namespace cuke::catalog
 {
 std::string as_readable_text()
 {
+  cuke::registry().sort_steps_by_type();
+
   std::string result{"Step Definitions (catalog):\n"};
   result.append("---------------------------\n");
   for (const auto& step : cuke::registry().steps())
@@ -22,7 +28,10 @@ std::string as_readable_text()
 
 std::string as_json(std::size_t indents /* = 2 */)
 {
+#ifdef HAS_JSON
   using json = nlohmann::json;
+
+  cuke::registry().sort_steps_by_type();
 
   json steps_catalog = {{"steps_catalog", json::array()}};
 
@@ -45,6 +54,21 @@ std::string as_json(std::size_t indents /* = 2 */)
   }
 
   return steps_catalog.dump(indents);
+#endif  // HAS_JSON
+
+  println(internal::color::red,
+          "nlohmann-json is not added as dependency in this build. Can not "
+          "export json format.");
+  return "";
+}
+
+void print_readable_text_to_sink()
+{
+  program_arguments().get_options().catalog.out.write(as_readable_text());
+}
+void print_json_to_sink(std::size_t indents /* = 2 */)
+{
+  program_arguments().get_options().catalog.out.write(as_json());
 }
 
 }  // namespace cuke::catalog