+++ src/cucumber.hpp
@@ -20,8 +20,11 @@ class cwt_cucumber
   cwt_cucumber(int argc, const char* argv[]);
   void run_tests() const noexcept;
   void print_results() const noexcept;
-  const options& get_options() const noexcept;
   [[nodiscard]] results::test_status final_result() const noexcept;
+
+  const options& get_options() const noexcept;
+  bool print_help() const noexcept;
+  bool export_catalog() const noexcept;
 };
 
 }  // namespace cuke